UInt32 hashes instead of UInt64

+++ src/Resin/DocHash.cs
@@ -4,15 +4,15 @@ namespace Resin
 {
     public class DocHash
     {
-        public UInt64 Hash { get; private set; }
+        public UInt32 Hash { get; private set; }
         public bool IsObsolete { get; set; }
 
-        public DocHash(UInt64 hash)
+        public DocHash(UInt32 hash)
         {
             Hash = hash;
         }
 
-        public DocHash(UInt64 hash, bool isObsolete)
+        public DocHash(UInt32 hash, bool isObsolete)
         {
             Hash = hash;
             IsObsolete = isObsolete;
